test(validation): NCO and LO after the control-port widening
Regenerated, and the verdicts move with the code rather than being left to
describe a tree that no longer exists.

NCO: F2 GAP -> FIXED (the port is double; delta 0), F3 GAP -> BY DESIGN (the
plateau is one phase-word LSB, the quantization floor, where the 22.4 ppb it
used to measure was the port's precision and not the accumulator's), F5 GAP
-> BY DESIGN. Three gaps to one.

F5 is the interesting one, and the reclassification is the user's call, not
a softening. It read "4 nonsense frequency requests produce a silently
stopped oscillator" -- but a rate below one LSB is not nonsense, it is an
ordinary request the phase word cannot represent, and truncating it to 0 IS
the law this object already documents: at most one step low, never high. It
is the bottom of the range behaving like the rest of it. NaN and +-inf reach
the same 0 through the conversion's deliberate totality. Grouping a
representable-but-tiny rate with a non-number was the error in the finding,
not the behaviour.

LO: F2 -> FIXED (the stale "rounds, not truncates" comment is corrected),
F3 -> FIXED (the AVX-512 blocks are deleted), F5 -> FIXED and F6 -> BY
DESIGN with the NCO's. Four gaps to zero, 26/26 limits.

The two reports' measurement code changed with them, not just their prose:
the fold prediction no longer round-trips through float32, and the dead-zone
scan is now spanned by the phase-word LSB rather than the float32 quantum --
otherwise both would have been measuring the old port through the new one.
That is what caught the fold limit at 72/1201.

Also fixes a generator-versus-formatter loop the summary line had all along
and only revealed at zero gaps: an empty gap list left "confirmed defects: "
with a trailing space, which the end-of-line hook strips and the next
regeneration puts back. The existing rstrip guards the same class of bug one
line further down.

+++ src/doppler/tests/validation/_common/linear_loop.py
@@ -138,7 +138,7 @@ def run(
     err = np.empty(n)
     ctl = np.empty(n)
     pha = np.empty(n)
-    one = np.zeros(1, dtype=np.float32)
+    one = np.zeros(1, dtype=np.float64)
     for k in range(n):
         pha[k] = nco.phase / W
         err[k] = detector(float(drive[k]), float(pha[k]))
